API

API

Functions

DiskJockey.constants.fftspace — Method.

Oftentimes it is necessary to get a symmetric coordinate array that spans $N$ elements from -width to +width, but makes sure that the middle point lands on $0$. The indices go from $0$ to $N -1.$ linspace returns the end points inclusive, wheras we want to leave out the right endpoint, because we are sampling the function in a cyclic manner.

source

Index

DiskJockey.constants.fftspace